Input: gt811 - Remove auto-update mechanism and make driver buildable
At this stage, the driver seems to talk to the touch screen:
 * configuration
 * periodic polling of touch coordinates


I haven't tested it for actual functionality though. Needless to say,
the driver needs massive clean up. The driver doesn't appear to have
been designed with portability in mind (hardcoded references to platform
specific GPIO API's for instance).

Signed-off-by: Mischa Jonker <mjonker@synopsys.com>